+++ silo-core/contracts/liquidation/lib/PartialLiquidationExecLib.sol
@@ -120,7 +120,7 @@ library PartialLiquidationExecLib {
         unchecked  { sumOfCollateralAssets = _ltvData.borrowerCollateralAssets + _ltvData.borrowerProtectedAssets; }
 
         if (_ltvData.borrowerDebtAssets == 0 || _params.debtToCover == 0) return (0, 0);
-        if (sumOfCollateralAssets == 0) return (0, _params.debtToCover);
+        if (sumOfCollateralAssets == 0) return (0, _params.debtToCover); // TODO cap
 
         (
             uint256 sumOfBorrowerCollateralValue, uint256 totalBorrowerDebtValue, uint256 ltvBefore